requirements.txt
discord.py
openpyxl
requests
BeautifulSoup